runner: update available architectures
This commit adds an aarch64-40pa archiecture to the runner invocation.
The runner is responsible for setting up the vspace of each component,
using only the architecture as an argument. aarch64-40pa is simply
aarch64 with a 3-level vspace starting at Level 1. This is only the case
for specific processors running in hypervisor mode.

Change-Id: I95f3bf217addd0cb93c35f01a73befbea21a05ac